Soft delete 구현, 좋아요 기능 구현

+++ src/main/java/goorm_runner/backend/post/dto/PostReadResponse.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 import java.time.LocalDateTime;
 
 @Builder
-public record PostReadResponse(String categoryName, Long postId, String title, String content, Short likeCount,
+public record PostReadResponse(String categoryName, Long postId, String title, String content, int likeCount,
                                LocalDateTime createdAt, LocalDateTime updatedAt) {
 }
